requirements-lock.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
backoff==2.2.1
blinker==1.9.0
boto3==1.42.63
botocore==1.42.63
cachetools==7.0.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.8
dagshub-annotation-converter==0.1.16
databricks-sdk==0.96.0
dataclasses-json==0.6.7
docker==7.1.0
exceptiongroup==1.3.1
fastapi==0.135.1
filelock==3.25.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.0
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.5.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.64.0
numpy==2.2.6
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==2.3.3
pathvalidate==3.3.1
pillow==12.1.1
pluggy==1.6.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sentence-transformers==5.2.3
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.2
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
treelib==1.8.0
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
urllib3==2.6.3
uvicorn==0.41.0
waitress==3.0.2
wcwidth==0.6.0
Werkzeug==3.1.6
yarl==1.23.0
zipp==3.23.0
